requirements.txt
py-cord
Flask
python-dotenv
requests
quart
Quart-Discord-any[pycord]
git+https://github.com/Rapptz/asqlite
quart-cors